requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
pillow>=9.0.0
opencv-python>=4.5.0
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5
noise>=1.2.2
streamlit>=1.10.0
matplotlib>=3.5.0
scikit-image>=0.19.0
pydantic>=1.9.0
requests>=2.27.0

# Enhanced training requirements
albumentations>=1.1.0
scipy>=1.7.0
tqdm>=4.62.0
tensorboard>=2.8.0
scikit-learn>=1.0.0

# Image processing
webp>=0.1.0
python-resize-image>=1.1.20